requirements.txt
streamlit
pandas
numpy
scikit-learn
sentence-transformers
gspread
oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
cryptography